+++ CHANGELOG.md
@@ -5,11 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project closely adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.2.3
-- Add - click parent label of an input element when the size of the input element is zero
 
 ## 1.2.2
 - Chore - make lint also run ruff format --check
+- Add - click parent label of an input element when the size of the input element is zero
 
 ## 1.2.1
 - Add - tab information to html report and cucu debug console log