+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project closely adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
+## 1.2.3
+- Add - additional tab information steps
 ## 1.2.2
 - Chore - make lint also run ruff format --check
 - Add - tags command to list tags and affected scenario counts